fix: useSearch 훅 사용으로 경로조회 페이지 Suspense 처리 (#212)
